+++ .env
@@ -1,4 +1,3 @@
 API_KEY_FMP='8fcb0de28daf241dd4522f0f455f2636'
-#API_KEY_FMP='gU76vxgpXMOG9nAkVumBE6bkruHKTgPb'
 API_KEY_FRED='bc393017b5507c40eca52633cdf37c93'
