+++ .idea/.name
@@ -1,1 +1,1 @@
-EDA & Preprocessing.py
+dividend_policy_predictor.py
