+++ dividend_policy_predictor.py
@@ -1,16 +1,3 @@
-# -*- coding: utf-8 -*-
-"""Dividend Policy Predictor
-
-Automatically generated by Colab.
-
-Original file is located at
-    https://colab.research.google.com/drive/1biyXJJEWsIIO2xEj79F4Lh-I5RvEOO3P
-"""
-
-! pip install python-dotenv
-
-! pip install optuna
-
 import pandas as pd
 import numpy as np
 import os
@@ -45,21 +32,21 @@ from xgboost import XGBClassifier
 warnings.filterwarnings('ignore')
 
 # Load Data
-dataset = pd.read_csv("Stock_data.csv")
+dataset = pd.read_csv("storage_files/Stock_data.csv")
 
 # Null value analysis
-dataset.info(verbose=True)
+print('Import original dataset.')
+# dataset.info(verbose=True)
 
 dataset.isna().sum()
 
-df = dataset #create copy df before one hot encoding
 enc = OrdinalEncoder()
-df = dataset.copy()
-df[["industry","sector", "symbol"]] = enc.fit_transform(df[["industry","sector", "symbol"]])
+df = dataset.copy() #create copy df before one hot encoding
+df[["industry","sector","symbol"]] = enc.fit_transform(df[["industry","sector","symbol"]])
 df.head()
 
 # Multivariate Analysis
-#target = dataset["dps_change_next_year"]
+# target = dataset["dps_change_next_year"]
 df.drop("dps_change_next_year", axis="columns", inplace=True)
 # Correlation matrix
 correlation_matrix = df.corr()
@@ -86,11 +73,12 @@ top_correlations = rank_columns_by_correlation(df, 0.98)
 # Remove highly correlated columns
 columns_to_remove = top_correlations["Column2"].unique()
 dataset.drop(columns_to_remove, axis="columns", inplace=True)
+print('Categorical variables:')
 print(columns_to_remove)
 
 # Data Preprocessing
 # Missing value
-dataset.info(verbose=True)
+# dataset.info(verbose=True)
 dataset.dropna(inplace=True)
 dataset.head()
 
@@ -98,6 +86,7 @@ dataset.head()
 training_data = dataset.loc[(dataset["year"] != 2022) & (dataset["year"] != 2021)]
 validation_data = dataset.loc[dataset["year"] == 2021]
 testing_data = dataset.loc[dataset["year"] == 2022]
+print('Target split finished.')
 
 # Predictor - Target Split
 y_train = training_data["dps_change_next_year"]
@@ -108,7 +97,7 @@ y_validate = validation_data["dps_change_next_year"]
 X_validate = validation_data.drop("dps_change_next_year", axis="columns")
 
 # Label encode categorical features with many categories
-categorical_columns = ["industry", "sector", "symbol"]
+categorical_columns = ["industry","sector","symbol"]
 other_columns = [col for col in X_train.columns if col not in categorical_columns]
 
 # Column Transformer
@@ -129,7 +118,8 @@ X_validate_transformed = pd.DataFrame(X_validate_transformed, columns=categorica
 X_test_transformed = pd.DataFrame(X_test_transformed, columns=categorical_columns + other_columns)
 
 # Check our data type
-X_train_transformed.info(verbose=True)
+# X_train_transformed.info(verbose=True)
+print('Categorical columns transformed.')
 
 # Let's change our data types back to their original forms - However, this time, categorical variables have become
 # number like strings
@@ -151,10 +141,11 @@ smote = SMOTENC(random_state=1, categorical_features=categorical_indices)
 X_train_oversampled, y_train_oversampled = smote.fit_resample(X_train_transformed, y_train)
 # Check our training data
 pd.DataFrame(y_train_oversampled)["dps_change_next_year"].value_counts()
-X_train_oversampled.info()
+# X_train_oversampled.info()
+print('Oversampling finished.')
 
 # Feature selection
-
+print('Feature selection started.')
 # Feature importance analysis - Tree Based
 randomForestModel = RandomForestClassifier(max_features=None)  # We want all features to be considered for each tree
 randomForestModel.fit(X_train_oversampled, y_train_oversampled)
@@ -174,6 +165,9 @@ plt.title("Feature Importance")
 plt.subplots_adjust(bottom=0.2, top=0.95)
 plt.xticks(rotation=45, ha='right', rotation_mode="anchor")
 plt.show()
+
+# ============================================================================
+
 
 # Now let's remove the features one by one from the least important one
 X_train_temp = X_train_oversampled.copy()
@@ -400,7 +394,7 @@ label_encoder = LabelEncoder()
 y_train_oversampled_encoded = label_encoder.fit_transform(y_train_oversampled)
 
 # Cast categorical types into numbers
-cols_to_convert = {'industry': 'float', 'sector': 'float', 'symbol': 'float'}
+cols_to_convert = {'industry': 'float', 'sector':'float', 'symbol': 'float'}
 X_train_oversampled = X_train_oversampled.astype(cols_to_convert)
 X_test_transformed = X_test_transformed.astype(cols_to_convert)
 # This suppresses printing logs
@@ -453,14 +447,14 @@ with open('best_models_xgb.pkl', 'wb') as file:
     pickle.dump(best_model_xgb, file)
 
 # Label encode categorical features with many categories
-column_transformer = ColumnTransformer(
-    transformers=[
-        ('categorical', OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1), categorical_columns)
-    ],
-    remainder='passthrough'
-)
-y = tmp[]
-X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
+# column_transformer = ColumnTransformer(
+#     transformers=[
+#         ('categorical', OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1), categorical_columns)
+#     ],
+#     remainder='passthrough'
+# )
+# y = tmp[]
+# X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
 
 # Model selection - Compare Performance
 with open('best_models_lr.pkl', 'rb') as file:
