+++ storage_files/importance_table.csv
@@ -1,0 +1,93 @@
+Feature,Importance
+dps_growth_rate,0.28673724631722963
+dps_growth,0.13528581941149365
+interestRate,0.08897539411182426
+year,0.06849185835734702
+dividendYield,0.061075933516740566
+interestRate_percentage_change,0.03591033614233716
+freeCashFlowPerShare,0.02865362493580012
+adjDividend,0.0207103457141221
+debtRatio,0.011537716187721345
+sector,0.011187332930976446
+assetTurnover,0.009636462054438051
+symbol,0.008030201251547263
+capitalExpenditureCoverageRatio,0.007969118421215471
+ebtPerEbit_percentage_change,0.005494757045714592
+priceToFreeCashFlowsRatio_percentage_change,0.005288928801042224
+operatingCycle_percentage_change,0.004840080105908747
+operatingCashFlowSalesRatio,0.004767345445054954
+dividendYield_percentage_change,0.004672428647715232
+industry,0.004610393396954005
+priceToSalesRatio,0.0045459030037701695
+enterpriseValueMultiple_percentage_change,0.004442021676852015
+operatingCashFlowPerShare,0.004203146354367
+daysOfPayablesOutstanding,0.004178311803732494
+netIncomePerEBT_percentage_change,0.0041183332125498085
+interestCoverage_percentage_change,0.003972928106071165
+priceToFreeCashFlowsRatio,0.003959138364535318
+priceEarningsToGrowthRatio_percentage_change,0.0038637046710615634
+grossProfitMargin_percentage_change,0.003862160722975417
+daysOfInventoryOutstanding,0.0038474210640540548
+fixedAssetTurnover,0.003833259215724571
+payablesTurnover,0.003765139279702885
+dividendPaidAndCapexCoverageRatio,0.0037122371354497445
+returnOnAssets,0.0036513828953107508
+grossProfitMargin,0.003649128426532133
+shortTermCoverageRatios_percentage_change,0.0034926291291670114
+daysOfSalesOutstanding,0.003438337204905513
+cashPerShare_percentage_change,0.0034274258217864497
+dividendPaidAndCapexCoverageRatio_percentage_change,0.0032798185808787877
+inventoryTurnover,0.0032477622754830907
+operatingProfitMargin,0.0030645448949258684
+quickRatio,0.003055591412139208
+netIncomePerEBT,0.003001935494302273
+freeCashFlowOperatingCashFlowRatio_percentage_change,0.002912889541606579
+effectiveTaxRate_percentage_change,0.002911341884645307
+payablesTurnover_percentage_change,0.0028894848420330664
+ebtPerEbit,0.0028547429014810916
+receivablesTurnover,0.00284997251063982
+payoutRatio,0.00284210544518068
+shortTermCoverageRatios,0.002825247602462103
+daysOfSalesOutstanding_percentage_change,0.002755690616876496
+longTermDebtToCapitalization,0.0027326742624665615
+payoutRatio_percentage_change,0.0027272379347671465
+priceToOperatingCashFlowsRatio,0.0027191228683791837
+returnOnEquity,0.0026164628605124325
+longTermDebtToCapitalization_percentage_change,0.0026009342582190677
+inventoryTurnover_percentage_change,0.0025639968569771635
+debtEquityRatio,0.002554802188008823
+freeCashFlowOperatingCashFlowRatio,0.0025450170814579186
+priceEarningsRatio,0.002537532823621844
+priceEarningsToGrowthRatio,0.002537485437052467
+cashFlowToDebtRatio,0.0024963294390596656
+priceBookValueRatio,0.002483402129050116
+freeCashFlowPerShare_percentage_change,0.0024650369918465215
+netProfitMargin,0.0023958490373383514
+interestCoverage,0.002385298955148729
+priceEarningsRatio_percentage_change,0.0023225987115168506
+returnOnAssets_percentage_change,0.0023132181404386483
+operatingCashFlowPerShare_percentage_change,0.0022407301646966537
+enterpriseValueMultiple,0.0022264167379542872
+companyEquityMultiplier_percentage_change,0.0022250740540578125
+companyEquityMultiplier,0.0022220881993092346
+priceToSalesRatio_percentage_change,0.0021881837233665453
+currentRatio,0.00216993514099443
+cashPerShare,0.0021059922584760974
+fixedAssetTurnover_percentage_change,0.0019776634851926124
+debtRatio_percentage_change,0.001976058838782599
+returnOnCapitalEmployed,0.0019301727110548508
+pretaxProfitMargin,0.0018870705143853883
+daysOfInventoryOutstanding_percentage_change,0.0018793847300247956
+assetTurnover_percentage_change,0.001844053949937242
+currentRatio_percentage_change,0.0017192192803384466
+totalDebtToCapitalization,0.0017021016585084118
+returnOnEquity_percentage_change,0.0016970480839978929
+capitalExpenditureCoverageRatio_percentage_change,0.0016598727233687138
+receivablesTurnover_percentage_change,0.0016265280098533264
+netProfitMargin_percentage_change,0.0015881280157941913
+priceToOperatingCashFlowsRatio_percentage_change,0.0015725661804054764
+cashFlowToDebtRatio_percentage_change,0.001459103573283767
+operatingProfitMargin_percentage_change,0.001239055117970412
+effectiveTaxRate,0.0012264555203541378
+debtEquityRatio_percentage_change,0.0011745922695566258
+pretaxProfitMargin_percentage_change,0.0011364461960894454
