+++ storage_files/result_df.csv
@@ -1,0 +1,93 @@
+Features_Removed,no_features_used,ROC_Score
+None,92,0.7359890109890109
+pretaxProfitMargin_percentage_change,91,0.7729395604395605
+debtEquityRatio_percentage_change,90,0.7572344322344322
+effectiveTaxRate,89,0.7139652014652016
+operatingProfitMargin_percentage_change,88,0.7190018315018315
+cashFlowToDebtRatio_percentage_change,87,0.7364468864468865
+priceToOperatingCashFlowsRatio_percentage_change,86,0.7312271062271063
+netProfitMargin_percentage_change,85,0.7313644688644688
+receivablesTurnover_percentage_change,84,0.7323717948717948
+capitalExpenditureCoverageRatio_percentage_change,83,0.7452380952380953
+returnOnEquity_percentage_change,82,0.7571886446886447
+totalDebtToCapitalization,81,0.7570512820512821
+currentRatio_percentage_change,80,0.7378205128205129
+assetTurnover_percentage_change,79,0.7670787545787545
+daysOfInventoryOutstanding_percentage_change,78,0.7471611721611722
+pretaxProfitMargin,77,0.7534798534798535
+returnOnCapitalEmployed,76,0.7673534798534799
+debtRatio_percentage_change,75,0.7697344322344323
+fixedAssetTurnover_percentage_change,74,0.7413919413919413
+cashPerShare,73,0.7549450549450549
+currentRatio,72,0.7477106227106226
+priceToSalesRatio_percentage_change,71,0.7076007326007326
+companyEquityMultiplier,70,0.7322344322344323
+companyEquityMultiplier_percentage_change,69,0.7593406593406593
+enterpriseValueMultiple,68,0.7249084249084249
+operatingCashFlowPerShare_percentage_change,67,0.7617216117216117
+returnOnAssets_percentage_change,66,0.7673534798534798
+priceEarningsRatio_percentage_change,65,0.771978021978022
+interestCoverage,64,0.7214285714285714
+netProfitMargin,63,0.7382326007326008
+freeCashFlowPerShare_percentage_change,62,0.7515567765567767
+priceBookValueRatio,61,0.7466575091575092
+cashFlowToDebtRatio,60,0.7618589743589743
+priceEarningsToGrowthRatio,59,0.7732142857142856
+priceEarningsRatio,58,0.7529761904761906
+freeCashFlowOperatingCashFlowRatio,57,0.7351648351648352
+debtEquityRatio,56,0.756364468864469
+inventoryTurnover_percentage_change,55,0.7570970695970697
+longTermDebtToCapitalization_percentage_change,54,0.7477106227106227
+returnOnEquity,53,0.777014652014652
+priceToOperatingCashFlowsRatio,52,0.7682692307692308
+payoutRatio_percentage_change,51,0.7653388278388279
+longTermDebtToCapitalization,50,0.7443223443223443
+daysOfSalesOutstanding_percentage_change,49,0.7629578754578754
+shortTermCoverageRatios,48,0.7696886446886447
+payoutRatio,47,0.7533424908424908
+receivablesTurnover,46,0.81753663003663
+ebtPerEbit,45,0.8005952380952381
+payablesTurnover_percentage_change,44,0.8079212454212454
+effectiveTaxRate_percentage_change,43,0.8159798534798535
+freeCashFlowOperatingCashFlowRatio_percentage_change,42,0.7576923076923077
+netIncomePerEBT,41,0.7802197802197802
+quickRatio,40,0.7991300366300366
+operatingProfitMargin,39,0.8007326007326007
+inventoryTurnover,38,0.7758241758241757
+dividendPaidAndCapexCoverageRatio_percentage_change,37,0.7972527472527472
+cashPerShare_percentage_change,36,0.8312728937728937
+daysOfSalesOutstanding,35,0.7567307692307692
+shortTermCoverageRatios_percentage_change,34,0.7941849816849818
+grossProfitMargin,33,0.7564102564102565
+returnOnAssets,32,0.8097527472527473
+dividendPaidAndCapexCoverageRatio,31,0.779532967032967
+payablesTurnover,30,0.759065934065934
+fixedAssetTurnover,29,0.7906593406593406
+daysOfInventoryOutstanding,28,0.7599358974358974
+grossProfitMargin_percentage_change,27,0.7566391941391942
+priceEarningsToGrowthRatio_percentage_change,26,0.7619047619047619
+priceToFreeCashFlowsRatio,25,0.7828754578754579
+interestCoverage_percentage_change,24,0.7837454212454211
+netIncomePerEBT_percentage_change,23,0.7861263736263737
+daysOfPayablesOutstanding,22,0.8003663003663004
+operatingCashFlowPerShare,21,0.7879120879120879
+enterpriseValueMultiple_percentage_change,20,0.8060439560439561
+priceToSalesRatio,19,0.7920787545787545
+industry,18,0.7994505494505494
+dividendYield_percentage_change,17,0.8179029304029304
+operatingCashFlowSalesRatio,16,0.8171703296703298
+operatingCycle_percentage_change,15,0.7935897435897437
+priceToFreeCashFlowsRatio_percentage_change,14,0.8262820512820513
+ebtPerEbit_percentage_change,13,0.8076465201465202
+capitalExpenditureCoverageRatio,12,0.7407967032967032
+symbol,11,0.7027930402930403
+assetTurnover,10,0.7176739926739927
+sector,9,0.7187271062271061
+debtRatio,8,0.6842948717948718
+adjDividend,7,0.6129578754578755
+freeCashFlowPerShare,6,0.6671703296703297
+interestRate_percentage_change,5,0.6421245421245421
+dividendYield,4,0.6248626373626374
+year,3,0.6013736263736262
+interestRate,2,0.5380036630036631
+dps_growth,1,0.6287545787545789
